% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_conversion.R
\name{construct_custom_var_mapping_from_spec_table}
\alias{construct_custom_var_mapping_from_spec_table}
\title{Construct a mapping of grouped custom variables from a "spec table"}
\usage{
construct_custom_var_mapping_from_spec_table(spec_table)
}
\arguments{
\item{spec_table}{A tibble/data frame with (at minimum) \code{database_id},
\code{measure_id}, \code{field_id}, and \code{value_id}. A \code{value_group} column may be
present for grouped recodes. If no \code{value_group} column returns NULL,}
}
\value{
A named list which can be input as \code{custom} input for \code{\link[=fetch_table]{fetch_table()}}
}
\description{
Takes grouped recodes are supported via a \code{value_group} column. Where \code{value_group}
is present, the function constructs the \code{custom} mapping needed by
\code{\link[=extract_results]{extract_results()}} so returned item labels align with the grouped query.
}
