% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_conversion.R
\name{construct_mapping_for_indiv_var}
\alias{construct_mapping_for_indiv_var}
\title{Build a Stat-Xplore recode mapping for a single field}
\usage{
construct_mapping_for_indiv_var(.tbl)
}
\arguments{
\item{.tbl}{A "spec table" subset for a single \code{field_id}.}
}
\value{
A list compatible with the Stat-Xplore JSON spec format for one
field, with elements \code{map} and \code{total}.
}
\description{
Build a Stat-Xplore recode mapping for a single field
}
\keyword{internal}
