% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_conversion.R
\name{convert_spec_table_to_list}
\alias{convert_spec_table_to_list}
\title{Convert a tidy table to a list (which is exportable to a json)}
\usage{
convert_spec_table_to_list(.table)
}
\description{
Convert a tidy table to a list (which is exportable to a json)
}
