% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_conversion.R
\name{convert_table_to_json}
\alias{convert_table_to_json}
\title{Convert a tidy table to JSON (list form)}
\usage{
convert_table_to_json(.table)
}
\description{
Convert a tidy table to JSON (list form)
}
