% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_helpers.R
\name{correct_json_format}
\alias{correct_json_format}
\title{Correct JSON formatting quirks}
\usage{
correct_json_format(.input_file, .output_file, remove_total_rows = T)
}
\description{
Correct JSON formatting quirks
}
\keyword{internal}
