% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_api.R
\name{get_next_level_info}
\alias{get_next_level_info}
\title{Get next-level info below a schema URL}
\usage{
get_next_level_info(url, error_if_next_level_empty = TRUE)
}
\description{
Get next-level info below a schema URL
}
