% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_api.R
\name{get_schema_info}
\alias{get_schema_info}
\title{Fetch schema info (with pagination via Link header)}
\usage{
get_schema_info(url, pages = 10)
}
\arguments{
\item{url}{Starting URL to fetch.}

\item{pages}{Maximum pages to traverse.}
}
\value{
Tibble of schema info (flattened).
}
\description{
Fetch schema info (with pagination via Link header)
}
