% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labeling.R
\name{get_target_type_info_below}
\alias{get_target_type_info_below}
\title{Traverse down from a URL to collect target types}
\usage{
get_target_type_info_below(.url, .target_types, .max_iterations = 10)
}
\description{
Traverse down from a URL to collect target types
}
