% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_conversion.R
\name{put_grouped_variable_mapping_into_table}
\alias{put_grouped_variable_mapping_into_table}
\title{Expand a grouped mapping in list format into a tidy specification table or "spec table"}
\usage{
put_grouped_variable_mapping_into_table(.mapping, .fieldname)
}
\arguments{
\item{.mapping}{A list where each element is a character vector of value ids.}

\item{.fieldname}{Field id to attach to the output.}
}
\value{
A tibble with columns \code{value_group}, \code{value}, and \code{field}.
}
\description{
Expand a grouped mapping in list format into a tidy specification table or "spec table"
}
\keyword{internal}
