% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_api.R
\docType{data}
\name{SCHEMA_URL}
\alias{SCHEMA_URL}
\title{Stat-Xplore API schema endpoint}
\format{
An object of class \code{character} of length 1.
}
\usage{
SCHEMA_URL
}
\description{
Stat-Xplore API schema endpoint
}
\keyword{internal}
