% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_config.R
\name{set_statxplore_config}
\alias{set_statxplore_config}
\alias{get_api_key}
\alias{get_base_url}
\title{Set and get configuration}
\usage{
set_statxplore_config(
  api_key = NULL,
  base_url = "https://stat-xplore.dwp.gov.uk/webapi/rest/v1"
)

get_api_key()

get_base_url()
}
\arguments{
\item{api_key}{Character. API key for authenticated calls (or set env var \code{STATXPLORE_API_KEY}).}

\item{base_url}{Character. Base schema URL. Defaults to "https://stat-xplore.dwp.gov.uk/webapi/rest/v1".}
}
\value{
Invisibly returns a named list of the values set.
}
\description{
Defaults to the official Stat-Xplore base URL.
}
