% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_spec.R
\name{update_spec_table_to_latest_data}
\alias{update_spec_table_to_latest_data}
\title{Update a spec table to include newer dates}
\usage{
update_spec_table_to_latest_data(.input_tbl)
}
\description{
Update a spec table to include newer dates
}
