% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\docType{data}
\name{URL_INFO}
\alias{URL_INFO}
\title{Stat-Xplore API info endpoint}
\format{
An object of class \code{character} of length 1.
}
\usage{
URL_INFO
}
\description{
Stat-Xplore API info endpoint
}
\keyword{internal}
