% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\docType{data}
\name{URL_TABLE}
\alias{URL_TABLE}
\title{Stat-Xplore API table endpoint}
\format{
An object of class \code{character} of length 1.
}
\usage{
URL_TABLE
}
\description{
Stat-Xplore API table endpoint
}
\keyword{internal}
