% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labeling.R
\name{add_labels_and_locations_to_spec_table}
\alias{add_labels_and_locations_to_spec_table}
\title{Enrich a spec table with labels and locations}
\usage{
add_labels_and_locations_to_spec_table(.input_tbl, .db_tbl = NULL)
}
\description{
Enrich a spec table with labels and locations
}
