% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_and_labels.R
\name{convert_labels_table_to_spec_table}
\alias{convert_labels_table_to_spec_table}
\title{Convert labels table to spec table with IDs}
\usage{
convert_labels_table_to_spec_table(
  .input_tbl,
  .db_tbl = NULL,
  .max_distance = 1
)
}
\description{
Convert labels table to spec table with IDs
}
