% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema_api.R
\name{list_databases}
\alias{list_databases}
\title{List databases from the configured base URL}
\usage{
list_databases(base_path = get_base_url())
}
\description{
List databases from the configured base URL
}
