% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_helpers.R
\name{remove_brackets_between_database_and_measures}
\alias{remove_brackets_between_database_and_measures}
\title{Remove brackets between database and measures}
\usage{
remove_brackets_between_database_and_measures(.input_string)
}
\description{
Remove brackets between database and measures
}
\keyword{internal}
