% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statxplorerplus-package.R
\docType{package}
\name{statxplorerplus-package}
\alias{statxplorerplus}
\alias{statxplorerplus-package}
\title{statxplorerplus: Construct queries to download data from the Stat-Xplore API}
\description{
The statxplorerplus package provides a suite of functions for
constructing queries and downloading data
from the Department for Work and Pensions Stat-Xplore API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://eboatlat.github.io/statxplorerplus/}
  \item \url{https://github.com/eboatlat/statxplorerplus}
  \item Report bugs at \url{https://github.com/eboatlat/statxplorerplus/issues}
}

}
\author{
\strong{Maintainer}: Eduin Latimer \email{Eduin.Latimer@ifs.org.uk}

Authors:
\itemize{
  \item Oli Hawkins
}

}
\keyword{internal}
