% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_and_labels.R
\name{stringdist_left_join_with_checks}
\alias{stringdist_left_join_with_checks}
\title{Fuzzy left join with safety checks}
\usage{
stringdist_left_join_with_checks(
  .left_df,
  .right_df,
  .max_distance = 1,
  .by,
  .no_duplicates = TRUE,
  .name_of_merge = "key"
)
}
\description{
Fuzzy left join with safety checks
}
